requirements.txt
python-telegram-bot==13.13
requests==2.28.1